fixing navabr
