+++ client/src/Pages/Home.js
@@ -2,7 +2,7 @@ import React from "react";
 
 function Home() {
     return (
-        <div>it works yee yee</div>
+        <div></div>
     )
 }
