+++ client/src/components/Navigation/Navigation.js
@@ -1,83 +1,58 @@
-import React from 'react';
-import { Box } from '@smooth-ui/core-sc';
-import { Link } from 'react-router-dom';
-import { withRouter } from 'react-router-dom';
+import React from "react";
+import { Box } from "@smooth-ui/core-sc";
+import { Link } from "react-router-dom";
+import { withRouter } from "react-router-dom";
 
-import { Logo } from '../index';
-import * as Styled from './Navigation.style';
+import Logo from "../Logo/Logo";
+import * as Styled from "./Navigation.style";
 
 const Navigation = ({ location }) => (
-  <Box
-    width="100%"
-    maxWidth="450px"
-    position="fixed"
-    left={0}
-    zIndex={100}
-    as="nav"
-    role="navigation"
-  >
     <Box
-      display="flex"
-      justifyContent="space-between"
-      alignItems="center"
-      p="10px 0 10px 10px"
+      width="100%"
+      maxWidth="450px"
+      position="fixed"
+      left={0}
+      zIndex={100}
+      as="nav"
+      role="navigation"
     >
-      <Styled.LogoLink
-        to="/"
-        aria-label="Home page"
+      <Box
+        display="flex"
+        justifyContent="space-between"
+        alignItems="center"
+        p="10px 0 10px 10px"
       >
-        <span
-          className="logo-link_content"
-          tabIndex="-1"
-        >
-          <Logo />
-        </span>
-      </Styled.LogoLink>
-      <Styled.NavList
-        page={location.pathname}
-      >
-        <li>
-          <Link
-            to="/about"
-            aria-label="About page"
-          >
-            <span
-              className="nav-link_content"
-              tabIndex="-1"
-            >
-              About
-            </span>
-          </Link>
-        </li>
-        <li>
-          <Link
-            to="/trails"
-            aria-label="Trails page"
-          >
-            <span
-              className="nav-link_content"
-              tabIndex="-1"
-            >
-              Trails
-            </span>
-          </Link>
-        </li>
-        <li>
-          <Link
-            to="/map"
-            aria-label="Trails page"
-          >
-            <span
-              className="nav-link_content"
-              tabIndex="-1"
-            >
-              Map
-            </span>
-          </Link>
-        </li>
-      </Styled.NavList>
+        <Styled.LogoLink to="/" aria-label="Home page">
+          <span className="logo-link_content" tabIndex="-1">
+            <Logo />
+          </span>
+        </Styled.LogoLink>
+        <Styled.NavList page={location.pathname}>
+          <li>
+            <Link to="/about" aria-label="About page">
+              <span className="nav-link_content" tabIndex="-1">
+                About
+              </span>
+            </Link>
+          </li>
+          <li>
+            <Link to="/trails" aria-label="Trails page">
+              <span className="nav-link_content" tabIndex="-1">
+                Trails
+              </span>
+            </Link>
+          </li>
+          <li>
+            <Link to="/map" aria-label="Trails page">
+              <span className="nav-link_content" tabIndex="-1">
+                Map
+              </span>
+            </Link>
+          </li>
+        </Styled.NavList>
+      </Box>
     </Box>
-  </Box>
+  
 );
 
 export default withRouter(Navigation);
