+++ client/src/components/index.js
@@ -1,10 +1,1 @@
-export { default as App } from './App/App';
-export { default as ScrollToTop } from './ScrollToTop/ScrollToTop';
-export { default as Pages } from './Pages/Pages';
-export { default as Navigation } from './Navigation/Navigation';
-export { default as Home } from './Home/Home';
-export { default as About } from './About/About';
-export { default as Trails } from './Trails/Trails';
-export { default as Logo } from './Logo/Logo';
-export { default as Facebook } from './Social/Facebook';
-export { default as Map } from './Map';
+export { default as App } from '../App';
